pyproject.toml
[tool.poetry]
name = "virus_q_api"
version = "1.0.0"
description = "Virus_Q API"
authors = ["ggidinson <dev@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.30"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
alembic-postgresql-enum = "^1.2.0"
aiohttp = "^3.9.5"
dishka = "^1.1.1"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
apscheduler = "^3.10.4"
tenacity = "^8.3.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.4.0"}
structlog = "^24.2.0"
fastapi-jwt = {extras = ["authlib"], version = "^0.3.0"}
sqlalchemy-utils = "^0.41.2"
granian = "^1.4.4"
gunicorn = "^22.0.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"
types-toml = "^0.10.8.20240310"
asyncpg-stubs = "^0.29.1"
types-pytz = "^2024.1.0.20240417"
types-sqlalchemy-utils = "^1.1.0"

[tool.ruff]
line-length = 100

[tool.ruff.format]
line-ending = "cr-lf"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "ISC", "Q", "LOG", "N", "C", "T20", "UP"]
ignore = ["B904", "B008", "ISC001", "N999"]

[[tool.mypy.overrides]]
module = [
    "migrations.versions.*"
]
disable_error_code = ["attr-defined", "arg-type"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
